Update : HistoryService 예외처리 & return 수정

+++ src/main/java/in/koala/serviceImpl/HistoryServiceImpl.java
@@ -1,6 +1,8 @@
 package in.koala.serviceImpl;
 
 import in.koala.domain.Notice;
+import in.koala.enums.ErrorMessage;
+import in.koala.exception.HistoryException;
 import in.koala.mapper.HistoryMapper;
 import in.koala.service.HistoryService;
 import in.koala.service.UserService;
@@ -19,7 +21,7 @@ public class HistoryServiceImpl implements HistoryService {
     private final UserService userService;
 
     @Override
-    public List<Notice> getEveryNotice(int pageNum) {
+    public List<Notice> getEveryNotice(int pageNum) throws Exception{
 
         Long userId = userService.getLoginUserInfo().getId();
 
@@ -28,22 +30,26 @@ public class HistoryServiceImpl implements HistoryService {
             pageNum = (pageNum * 10) + 1;
         }
 
-        return historyMapper.getEveryNotice(userId, pageNum);
+        List<Notice> result = historyMapper.getEveryNotice(userId, pageNum);
+
+        if(result.isEmpty()){
+            throw new HistoryException(ErrorMessage.NOTICE_NOT_EXIST);
+        }
+        else
+            return result;
     }
 
     @Override
     public void deleteNotice(List<Integer> noticeList) {
 
+        if(noticeList.isEmpty())
+            throw new HistoryException(ErrorMessage.NOTICE_NOT_SELECTED);
+
         historyMapper.deleteNotice(noticeList);
     }
 
     @Override
-    public Boolean noticeRead(String noticeId) {
-        if(historyMapper.noticeRead(noticeId) == 1){
-            return true;
-        }
-        else{
-            return false;
-        }
+    public void noticeRead(String noticeId) {
+        historyMapper.noticeRead(noticeId);
     }
 }
